Move from aiohttp_auth to aiohttp_auth_autz.
- Change name in setup.
- Remove original tests.
- Make pytests as default tests.
