+++ .gitignore
@@ -4,4 +4,5 @@ dist
 .coverage
 .tox
 *.egg
+*.eggs
 *.egg-info
