+++ setup.cfg
@@ -1,0 +1,4 @@
+[aliases]
+test=pytest
+[tool:pytest]
+addopts = -v --cov-report=term-missing --cov=aiohttp_auth --cov=tests
