+++ tox.ini
@@ -7,4 +7,4 @@ deps=
     pytest-aiohttp
     pytest-cov
     aiohttp_session
-commands=py.test -v --cov-report=term-missing --cov=aiohttp_auth --cov=pytests pytests
+commands=py.test -v --cov-report=term-missing --cov=aiohttp_auth --cov=tests tests
